requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Limiter==3.3.1
python-dotenv==1.0.0
watchtower==3.4.0
Flask-Migrate==4.0.4
# Database adapters
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
# Image analysis
boto3>=1.28.0
Pillow==10.1.0
python-multipart==0.0.6
# Additional security
PyJWT==2.8.0
cryptography==41.0.4
